feat: 푸터 반응형 대응 (#262)
